pyproject.toml
[tool.poetry]
name = "streetwarp-lambda"
version = "0.1.0"
description = "AWS Lambda program which executes streetwarp"
authors = ["Peter Elmers <peter.elmers@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.7"
websockets = "^8.1"
azure-storage-blob = "^12.8.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
